+++ library/src/actions/rfcEmail/rfcEmail.test.ts
@@ -80,15 +80,15 @@ describe('rfcEmail', () => {
 
     // Start of local part
 
-    test('for underscore in begining of local part', () => {
+    test('for underscore in beginning of local part', () => {
       expectNoActionIssue(action, ['_email@example.com']);
     });
 
-    test('for hyphen in begining of local part', () => {
+    test('for hyphen in beginning of local part', () => {
       expectNoActionIssue(action, ['-email@example.com']);
     });
 
-    test('for plus in begining of local part', () => {
+    test('for plus in beginning of local part', () => {
       expectNoActionIssue(action, ['+email@example.com']);
     });